Make GIN tests run explicitly (#310)
* re-organized test structure for workflow

* testing explicit failure

* modified workflows to explicitly trigger GIN

* modified workflows to explicitly trigger GIN

* fixing CI

* fixing test_gin location

* fixing partial test location

* fixed remove of checking functions

* fixed names; restored PR run

* restored PR for run

* Update update-testing-data.yml

* Update update-testing-data.yml

* test set local

* Automated changes

* fix set env

* print env

* print env

* Automated changes

* one more test

* Automated changes

* Update tests/test_on_data/test_gin.py

* cleaning up; adjusting cache keys

* Update update-testing-data.yml

* updating cache key
