--- .github/workflows/full-tests-mac-windows.yml
@@ -1,35 +1,0 @@
-name: Full tests on Mac and Windows
-on: pull_request
-jobs:
-  run:
-    name: Full tests on ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ["macos-latest", "windows-latest"]
-    steps:
-    - uses: actions/checkout@v2
-    - run: git fetch --prune --unshallow --tags
-    - name: Setup Python
-      uses: actions/setup-python@v2
-      with:
-        python-version: 3.7
-    - name: Install this package and pytest, on development mode
-      env:
-          NWB_CONVERSION_INSTALL_MODE: development
-      run: |
-        pip install -r requirements-full.txt
-        pip install -e .
-        pip install pytest
-    - name: Get ephy_testing_data current head hash
-      id: vars
-      run: echo "::set-output name=HASH_EPHY_DATASET::$(git ls-remote https://gin.g-node.org/NeuralEnsemble/ephy_testing_data.git HEAD | cut -f1)"
-    - uses: actions/cache@v2
-      id: cache-datasets
-      with:
-        path: /home/runner/work/nwb-conversion-tools/nwb-conversion-tools/ephy_testing_data
-        key: ecephys-dset-${{ steps.vars.outputs.HASH_EPHY_DATASET }}
-        restore-keys: ecephys-dset-${{ steps.vars.outputs.HASH_EPHY_DATASET }}
-    - name: Run pytest
-      run: pytest
