+++ .github/workflows/update-testing-data.yml
@@ -29,15 +29,15 @@ jobs:
         id: cache-ephys-dataset
         with:
           path: /home/runner/work/nwb-conversion-tools/nwb-conversion-tools/ephy_testing_data
-          key: ephys-datasets-${{ steps.ephys.outputs.HASH_EPHYS_DATASET }}
-          restore-keys: ephys-datasets-${{ steps.ephys.outputs.HASH_EPHYS_DATASET }}
+          key: ephys-datasets-1-${{ steps.ephys.outputs.HASH_EPHYS_DATASET }}
+          restore-keys: ephys-datasets-1-${{ steps.ephys.outputs.HASH_EPHYS_DATASET }}
       - name: Cache ophys dataset - ${{ steps.ophys.outputs.HASH_OPHYS_DATASET }}
         uses: actions/cache@v2
         id: cache-ophys-dataset
         with:
           path: /home/runner/work/nwb-conversion-tools/nwb-conversion-tools/ophys_testing_data
-          key: ophys-datasets-${{ steps.ophys.outputs.HASH_OPHYS_DATASET }}
-          restore-keys: ophys-datasets-${{ steps.ophys.outputs.HASH_OPHYS_DATASET }}
+          key: ophys-datasets-1-${{ steps.ophys.outputs.HASH_OPHYS_DATASET }}
+          restore-keys: ophys-datasets-1-${{ steps.ophys.outputs.HASH_OPHYS_DATASET }}
       - name: Update datasets
         run: |
           datalad install -r -g https://gin.g-node.org/NeuralEnsemble/ephy_testing_data
